requirements.txt
django>=4.2,<5.0
pillow>=10.0.0
pillow-heif>=0.13.0
requests>=2.31.0


